Follow system link
Added functionality to the script to enable its usage via bash from any part
of the filesystem by creating a symlink of the main script.
Implemented File.realpath to properly follow the __FILE__ in case it is a
symlink. This addition does not interfere with the correct behavior of the
script.